Conflicts:
	src/Illuminate/Database/MigrationServiceProvider.php

+++ src/Illuminate/Database/Eloquent/Relations/BelongsToMany.php
@@ -1,9 +1,9 @@
 <?php namespace Illuminate\Database\Eloquent\Relations;
 
-use DateTime;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class BelongsToMany extends Relation {
 
@@ -74,6 +74,34 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
+	 * Set a where clause for a pivot table column.
+	 *
+	 * @param  string  $column
+	 * @param  string  $operator
+	 * @param  mixed   $value
+	 * @param  string  $boolean
+	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
+	 */
+	public function wherePivot($column, $operator = null, $value = null, $boolean = 'and')
+	{
+		return $this->where($this->table.'.'.$column, $operator, $value, $boolean);
+	}
+
+	/**
+	 * Set an or where clause for a pivot table column.
+	 *
+	 * @param  string  $column
+	 * @param  string  $operator
+	 * @param  mixed   $value
+	 * @param  string  $boolean
+	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
+	 */
+	public function orWherePivot($column, $operator = null, $value = null)
+	{
+		return $this->wherePivot($column, $operator, $value, 'or');
+	}
+
+	/**
 	 * Execute the query and get the first result.
 	 *
 	 * @param  array   $columns
@@ -87,6 +115,21 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
+	 * Execute the query and get the first result or throw an exception.
+	 *
+	 * @param  array  $columns
+	 * @return \Illuminate\Database\Eloquent\Model|static
+	 *
+	 * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
+	 */
+	public function firstOrFail($columns = array('*'))
+	{
+		if ( ! is_null($model = $this->first($columns))) return $model;
+
+		throw new ModelNotFoundException;
+	}
+
+	/**
 	 * Execute the query as a "select" statement.
 	 *
 	 * @param  array  $columns
@@ -94,7 +137,12 @@ class BelongsToMany extends Relation {
 	 */
 	public function get($columns = array('*'))
 	{
-		$models = $this->query->getModels($this->getSelectColumns($columns));
+		// First we'll add the proper select columns onto the query so it is run with
+		// the proper columns. Then, we will get the results and hydrate out pivot
+		// models with the result of those columns as a separate model relation.
+		$select = $this->getSelectColumns($columns);
+
+		$models = $this->query->addSelect($select)->getModels();
 
 		$this->hydratePivotRelation($models);
 
@@ -110,6 +158,27 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
+	 * Get a paginator for the "select" statement.
+	 *
+	 * @param  int    $perPage
+	 * @param  array  $columns
+	 * @return \Illuminate\Pagination\Paginator
+	 */
+	public function paginate($perPage = null, $columns = array('*'))
+	{
+		$this->query->addSelect($this->getSelectColumns($columns));
+
+		// When paginating results, we need to add the pivot columns to the query and
+		// then hydrate into the pivot objects once the results have been gathered
+		// from the database since this isn't performed by the Eloquent builder.
+		$pager = $this->query->paginate($perPage, $columns);
+
+		$this->hydratePivotRelation($pager->getItems());
+
+		return $pager;
+	}
+
+	/**
 	 * Hydrate the pivot table relationship on the models.
 	 *
 	 * @param  array  $models
@@ -122,9 +191,7 @@ class BelongsToMany extends Relation {
 		// will set the attributes, table, and connections on so it they be used.
 		foreach ($models as $model)
 		{
-			$values = $this->cleanPivotAttributes($model);
-
-			$pivot = $this->newExistingPivot($values);
+			$pivot = $this->newExistingPivot($this->cleanPivotAttributes($model));
 
 			$model->setRelation('pivot', $pivot);
 		}
@@ -163,13 +230,66 @@ class BelongsToMany extends Relation {
 	 */
 	public function addConstraints()
 	{
-		$this->setJoin()->setWhere();
+		$this->setJoin();
+
+		if (static::$constraints) $this->setWhere();
+	}
+
+	/**
+	 * Add the constraints for a relationship count query.
+	 *
+	 * @param  \Illuminate\Database\Eloquent\Builder  $query
+	 * @param  \Illuminate\Database\Eloquent\Builder  $parent
+	 * @return \Illuminate\Database\Eloquent\Builder
+	 */
+	public function getRelationCountQuery(Builder $query, Builder $parent)
+	{
+		if ($parent->getQuery()->from == $query->getQuery()->from)
+		{
+			return $this->getRelationCountQueryForSelfJoin($query, $parent);
+		}
+		else
+		{
+			$this->setJoin($query);
+
+			return parent::getRelationCountQuery($query, $parent);
+		}
+	}
+
+	/**
+	 * Add the constraints for a relationship count query on the same table.
+	 *
+	 * @param  \Illuminate\Database\Eloquent\Builder  $query
+	 * @param  \Illuminate\Database\Eloquent\Builder  $parent
+	 * @return \Illuminate\Database\Eloquent\Builder
+	 */
+	public function getRelationCountQueryForSelfJoin(Builder $query, Builder $parent)
+	{
+		$query->select(new \Illuminate\Database\Query\Expression('count(*)'));
+
+		$tablePrefix = $this->query->getQuery()->getConnection()->getTablePrefix();
+
+		$query->from($this->table.' as '.$tablePrefix.$hash = $this->getRelationCountHash());
+
+		$key = $this->wrap($this->getQualifiedParentKeyName());
+
+		return $query->where($hash.'.'.$this->foreignKey, '=', new \Illuminate\Database\Query\Expression($key));
+	}
+
+	/**
+	 * Get a relationship join table hash.
+	 *
+	 * @return string
+	 */
+	public function getRelationCountHash()
+	{
+		return 'self_'.md5(microtime(true));
 	}
 
 	/**
 	 * Set the select clause for the relation query.
 	 *
-	 * @return \Illuminate\Database\Eloquent\Relation\BelongsToMany
+	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
 	 */
 	protected function getSelectColumns(array $columns = array('*'))
 	{
@@ -206,10 +326,13 @@ class BelongsToMany extends Relation {
 	/**
 	 * Set the join clause for the relation query.
 	 *
-	 * @return \Illuminate\Database\Eloquent\Relation\BelongsToMany
+	 * @param  \Illuminate\Database\Eloquent\Builder|null
+	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
 	 */
-	protected function setJoin()
+	protected function setJoin($query = null)
 	{
+		$query = $query ?: $this->query;
+
 		// We need to join to the intermediate table on the related model's primary
 		// key column with the intermediate table's foreign key for the related
 		// model instance. Then we can set the "where" for the parent models.
@@ -217,7 +340,7 @@ class BelongsToMany extends Relation {
 
 		$key = $baseTable.'.'.$this->related->getKeyName();
 
-		$this->query->join($this->table, $key, '=', $this->getOtherKey());
+		$query->join($this->table, $key, '=', $this->getOtherKey());
 
 		return $this;
 	}
@@ -225,7 +348,7 @@ class BelongsToMany extends Relation {
 	/**
 	 * Set the where clause for the relation query.
 	 *
-	 * @return \Illuminate\Database\Eloquent\Relation\BelongsToMany
+	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
 	 */
 	protected function setWhere()
 	{
@@ -252,7 +375,7 @@ class BelongsToMany extends Relation {
 	 *
 	 * @param  array   $models
 	 * @param  string  $relation
-	 * @return void
+	 * @return array
 	 */
 	public function initRelation(array $models, $relation)
 	{
@@ -326,7 +449,7 @@ class BelongsToMany extends Relation {
 	{
 		$key = $this->getRelated()->getKeyName();
 
-		$columns = array($this->getRelatedUpdated() => new DateTime);
+		$columns = $this->getRelatedFreshUpdate();
 
 		// If we actually have IDs for the relation, we will run the query to update all
 		// the related model's timestamps, to make sure these all reflect the changes
@@ -414,8 +537,8 @@ class BelongsToMany extends Relation {
 	/**
 	 * Create an array of new instances of the related models.
 	 *
-	 * @param  array  $attributes
-	 * @param  array  $joining
+	 * @param  array  $records
+	 * @param  array  $joinings
 	 * @return \Illuminate\Database\Eloquent\Model
 	 */
 	public function createMany(array $records, array $joinings = array())
@@ -433,13 +556,20 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
-	 * Sync the intermediate tables with a list of IDs.
+	 * Sync the intermediate tables with a list of IDs or collection of models.
 	 *
-	 * @param  array  $ids
-	 * @return void
+	 * @param  $ids
+	 * @param  bool   $detaching
+	 * @return array
 	 */
-	public function sync(array $ids)
+	public function sync($ids, $detaching = true)
 	{
+		$changes = array(
+			'attached' => array(), 'detached' => array(), 'updated' => array()
+		);
+
+		if ($ids instanceof Collection) $ids = $ids->modelKeys();
+
 		// First we need to attach any of the associated models that are not currently
 		// in this joining table. We'll spin through the given IDs, checking to see
 		// if they exist in the array of current ones, and if not we will insert.
@@ -452,21 +582,30 @@ class BelongsToMany extends Relation {
 		// Next, we will take the differences of the currents and given IDs and detach
 		// all of the entities that exist in the "current" array but are not in the
 		// the array of the IDs given to the method which will complete the sync.
-		if (count($detach) > 0)
+		if ($detaching && count($detach) > 0)
 		{
 			$this->detach($detach);
+
+			$changes['detached'] = (array) array_map('intval', $detach);
 		}
 
 		// Now we are finally ready to attach the new records. Note that we'll disable
 		// touching until after the entire operation is complete so we don't fire a
 		// ton of touch operations until we are totally done syncing the records.
-		$this->attachNew($records, $current, false);
+		$changes = array_merge(
+			$changes, $this->attachNew($records, $current, false)
+		);
 
-		$this->touchIfTouching();
+		if (count($changes['attached']) || count($changes['updated']))
+		{
+			$this->touchIfTouching();
+		}
+
+		return $changes;
 	}
 
 	/**
-	 * Format the sync list so that is is keyed by ID.
+	 * Format the sync list so that it is keyed by ID.
 	 *
 	 * @param  array  $records
 	 * @return array
@@ -477,9 +616,9 @@ class BelongsToMany extends Relation {
 
 		foreach ($records as $id => $attributes)
 		{
-			if (is_numeric($attributes))
+			if ( ! is_array($attributes))
 			{
-				list($id, $attributes) = array((int) $attributes, array());
+				list($id, $attributes) = array($attributes, array());
 			}
 
 			$results[$id] = $attributes;
@@ -494,14 +633,58 @@ class BelongsToMany extends Relation {
 	 * @param  array  $records
 	 * @param  array  $current
 	 * @param  bool   $touch
-	 * @return void
+	 * @return array
 	 */
 	protected function attachNew(array $records, array $current, $touch = true)
 	{
+		$changes = array('attached' => array(), 'updated' => array());
+
 		foreach ($records as $id => $attributes)
 		{
-			if ( ! in_array($id, $current)) $this->attach($id, $attributes, $touch);
+			// If the ID is not in the list of existing pivot IDs, we will insert a new pivot
+			// record, otherwise, we will just update this existing record on this joining
+			// table, so that the developers will easily update these records pain free.
+			if ( ! in_array($id, $current))
+			{
+				$this->attach($id, $attributes, $touch);
+
+				$changes['attached'][] = (int) $id;
+			}
+
+			// Now we'll try to update an existing pivot record with the attributes that were
+			// given to the method. If the model is actually updated we will add it to the
+			// list of updated pivot records so we return them back out to the consumer.
+			elseif (count($attributes) > 0)
+			{
+				if ($this->updateExistingPivot($id, $attributes, $touch))
+				{
+					$changes['updated'][] = (int) $id;
+				}
+			}
 		}
+		return $changes;
+	}
+
+	/**
+	 * Update an existing pivot record on the table.
+	 *
+	 * @param  mixed  $id
+	 * @param  array  $attributes
+	 * @param  bool   $touch
+	 * @return void
+	 */
+	public function updateExistingPivot($id, array $attributes, $touch = true)
+	{
+		if (in_array($this->updatedAt(), $this->pivotColumns))
+		{
+			$attributes = $this->setTimestampsOnAttach($attributes, true);
+		}
+
+		$updated = $this->newPivotStatementForId($id)->update($attributes);
+
+		if ($touch) $this->touchIfTouching();
+
+		return $updated;
 	}
 
 	/**
@@ -527,7 +710,7 @@ class BelongsToMany extends Relation {
 	 * Create an array of records to insert into the pivot table.
 	 *
 	 * @param  array  $ids
-	 * @return void
+	 * @return array
 	 */
 	protected function createAttachRecords($ids, array $attributes)
 	{
@@ -605,10 +788,26 @@ class BelongsToMany extends Relation {
 		// provide us with a fresh timestamp in this model's preferred format.
 		if ($timed)
 		{
-			$record[$this->createdAt()] = $this->parent->freshTimestamp();
-
-			$record[$this->updatedAt()] = $record[$this->createdAt()];
+			$record = $this->setTimestampsOnAttach($record);
 		}
+
+		return $record;
+	}
+
+	/**
+	 * Set the creation and update timestamps on an attach record.
+	 *
+	 * @param  array  $record
+	 * @param  bool   $exists
+	 * @return array
+	 */
+	protected function setTimestampsOnAttach(array $record, $exists = false)
+	{
+		$fresh = $this->parent->freshTimestamp();
+
+		if ( ! $exists) $record[$this->createdAt()] = $fresh;
+
+		$record[$this->updatedAt()] = $fresh;
 
 		return $record;
 	}
@@ -652,10 +851,10 @@ class BelongsToMany extends Relation {
 	 * @return void
 	 */
 	public function touchIfTouching()
-	{ 
-	 	if ($this->touchingParent()) $this->getParent()->touch();
+	{
+		if ($this->touchingParent()) $this->getParent()->touch();
 
-	 	if ($this->getParent()->touches($this->relationName)) $this->touch();
+		if ($this->getParent()->touches($this->relationName)) $this->touch();
 	}
 
 	/**
@@ -675,7 +874,7 @@ class BelongsToMany extends Relation {
 	 */
 	protected function guessInverseRelation()
 	{
-		return strtolower(str_plural(class_basename($this->getParent())));
+		return camel_case(str_plural(class_basename($this->getParent())));
 	}
 
 	/**
@@ -685,7 +884,7 @@ class BelongsToMany extends Relation {
 	 */
 	protected function newPivotQuery()
 	{
-		$query = $this->query->getQuery()->newQuery()->from($this->table);
+		$query = $this->newPivotStatement();
 
 		return $query->where($this->foreignKey, $this->parent->getKey());
 	}
@@ -701,19 +900,32 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
+	 * Get a new pivot statement for a given "other" ID.
+	 *
+	 * @param  mixed  $id
+	 * @return \Illuminate\Database\Query\Builder
+	 */
+	public function newPivotStatementForId($id)
+	{
+		$pivot = $this->newPivotStatement();
+
+		$key = $this->parent->getKey();
+
+		return $pivot->where($this->foreignKey, $key)->where($this->otherKey, $id);
+	}
+
+	/**
 	 * Create a new pivot model instance.
 	 *
 	 * @param  array  $attributes
 	 * @param  bool   $exists
-	 * @return \Illuminate\Database\Eloquent\Relation\Pivot
+	 * @return \Illuminate\Database\Eloquent\Relations\Pivot
 	 */
 	public function newPivot(array $attributes = array(), $exists = false)
 	{
-		$pivot = new Pivot($this->parent, $attributes, $this->table, $exists);
+		$pivot = $this->related->newPivot($this->parent, $attributes, $this->table, $exists);
 
-		$pivot->setPivotKeys($this->foreignKey, $this->otherKey);
-
-		return $pivot;
+		return $pivot->setPivotKeys($this->foreignKey, $this->otherKey);
 	}
 
 	/**
@@ -735,7 +947,9 @@ class BelongsToMany extends Relation {
 	 */
 	public function withPivot($columns)
 	{
-		$this->pivotColumns = is_array($columns) ? $columns : func_get_args();
+		$columns = is_array($columns) ? $columns : func_get_args();
+
+		$this->pivotColumns = array_merge($this->pivotColumns, $columns);
 
 		return $this;
 	}
@@ -745,13 +959,9 @@ class BelongsToMany extends Relation {
 	 *
 	 * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
 	 */
-	public function withTimestamps()
+	public function withTimestamps($createdAt = null, $updatedAt = null)
 	{
-		$columns = array($this->createdAt(), $this->updatedAt());
-
-		$this->pivotColumns = array_merge($this->pivotColumns, $columns);
-
-		return $this;
+		return $this->withPivot($createdAt ?: $this->createdAt(), $updatedAt ?: $this->updatedAt());
 	}
 
 	/**
@@ -759,9 +969,19 @@ class BelongsToMany extends Relation {
 	 *
 	 * @return string
 	 */
-	public function getRelatedUpdated()
+	public function getRelatedFreshUpdate()
 	{
-		return $this->getRelated()->getUpdatedAtColumn();
+		return array($this->related->getUpdatedAtColumn() => $this->related->freshTimestamp());
+	}
+
+	/**
+	 * Get the key for comparing against the parent key in "has" query.
+	 *
+	 * @return string
+	 */
+	public function getHasCompareKey()
+	{
+		return $this->getForeignKey();
 	}
 
 	/**
@@ -785,6 +1005,16 @@ class BelongsToMany extends Relation {
 	}
 
 	/**
+	 * Get the fully qualified parent key name.
+	 *
+	 * @return string
+	 */
+	protected function getQualifiedParentKeyName()
+	{
+		return $this->parent->getQualifiedKeyName();
+	}
+
+	/**
 	 * Get the intermediate table for the relationship.
 	 *
 	 * @return string
@@ -792,6 +1022,16 @@ class BelongsToMany extends Relation {
 	public function getTable()
 	{
 		return $this->table;
+	}
+
+	/**
+	 * Get the relationship name for the relationship.
+	 *
+	 * @return string
+	 */
+	public function getRelationName()
+	{
+		return $this->relationName;
 	}
 
 }
